refactor: clean change helper
